+++ packages/browser-wallet/src/popup/pages/Account/i18n/en.ts
@@ -2,6 +2,7 @@ const t = {
     noAccounts: 'No accounts in wallet',
     removeAccount: 'Remove account (local only)',
     resetConnections: 'Reset connections',
+    accountAddress: 'Account address',
     actions: {
         log: 'transaction log',
         send: 'send ccd',