+++ packages/browser-wallet/src/popup/pages/Account/i18n/da.ts
@@ -16,6 +16,15 @@ const t: typeof en = {
         atDisposal: 'Offentligt til rådighed',
         stakeAmount: 'Stake',
     },
+    settings: {
+        connectedSites: {
+            title: 'Forbundne hjemmesider',
+            noConnected: 'Den valgte konto er ikke forbundet til nogen hjemmeside.',
+            connect: 'Forbind',
+            disconnect: 'Fjern',
+        },
+        exportPrivateKey: 'Eksportér privat nøgle',
+    },
 };
 
 export default t;