+++ packages/browser-wallet/CHANGELOG.md
@@ -6,6 +6,12 @@
 
 -   Added display of address as text and qr code in the "receive ccd" section.
 
+## 0.2.1
+
+### Fixed
+
+-   Sending V1 contracts updates/initializations with parameters now works.
+
 ## 0.2.0
 
 ### Added